Split permission rule safety architecture
