+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -12,7 +12,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `source:QuillCodeApp` | 446 | 50972 | 100 | A+ | `WorkspaceSwiftUIView.swift` A+<br>`QuillCodeTranscriptSurface.swift` A+<br>`QuillCodeReviewSurface.swift` A+ |
 | `source:QuillCodeCore` | 33 | 3199 | 99 | A+ | `AutomationModels.swift` A+<br>`AppConfig.swift` A+<br>`ModelInfo.swift` A+ |
 | `source:QuillCodePersistence` | 8 | 767 | 100 | A+ | `PermissionRuleStore.swift` A+<br>`ConfigStore.swift` A+<br>`SecretStore.swift` A+ |
-| `source:QuillCodeSafety` | 11 | 2087 | 99 | A+ | `PermissionRuleTable.swift` A+<br>`Safety.swift` A+<br>`PermissionRuleGatedSafetyReviewer.swift` A+ |
+| `source:QuillCodeSafety` | 13 | 2145 | 99 | A+ | `Safety.swift` A+<br>`PermissionRuleGatedSafetyReviewer.swift` A+<br>`StaticSafetyReadOnlyShellPolicy.swift` A+ |
 | `source:QuillCodeTools` | 95 | 10529 | 100 | A+ | `GitToolExecutor.swift` A+<br>`ShellToolExecutor.swift` A+<br>`GitPatchToolExecutor.swift` A+ |
 | `source:QuillComputerUseKit` | 11 | 1285 | 99 | A+ | `LinuxComputerUseBackend.swift` A+<br>`ComputerUseContracts.swift` A+<br>`LinuxComputerUseCapabilityDetector.swift` A+ |
 | `source:quill-code` | 1 | 171 | 99 | A+ | `main.swift` A+ |
@@ -32,7 +32,6 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 
 | Grade | Score | Lines | Module | File | Main Issues |
 | --- | ---: | ---: | --- | --- | --- |
-| A+ | 96 | 338 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleTable.swift` | 2 lines >120 chars; 14 public declarations |
 | A+ | 96 | 268 | `source:QuillCodeAgent` | `Sources/QuillCodeAgent/TrustedRouterLLMClient.swift` | 1 lines >120 chars; 14 public declarations |
 | A+ | 97 | 635 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySwiftInteractionTargetAuditSupport.swift` | 9 lines >120 chars; 1 lines >160 chars |
 | A+ | 97 | 614 | `source:quill-code-desktop` | `Sources/quill-code-desktop/QuillCodeDesktopController.swift` | watch file size (614 lines) |
@@ -72,6 +71,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 97 | 182 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeSidebarFilterSurface.swift` | 30 public declarations |
 | A+ | 97 | 181 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleGatedSafetyReviewer.swift` | 13 public declarations |
 | A+ | 97 | 172 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeMCPSurface.swift` | 26 public declarations |
+| A+ | 97 | 172 | `source:QuillCodeApp` | `Sources/QuillCodeApp/QuillCodeNativeHitTargetContract.swift` | 20 public declarations |
 
 ## Every File
 
@@ -703,8 +703,10 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 74 | `source:QuillCodePersistence` | `Sources/QuillCodePersistence/SecretStore.swift` | no automated issues |
 | A+ | 100 | 49 | `source:QuillCodePersistence` | `Sources/QuillCodePersistence/ThreadStore.swift` | no automated issues |
 | A+ | 97 | 181 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleGatedSafetyReviewer.swift` | 13 public declarations |
+| A+ | 100 | 94 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleModels.swift` | no automated issues |
 | A+ | 100 | 258 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleSubject.swift` | no automated issues |
-| A+ | 96 | 338 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleTable.swift` | 2 lines >120 chars; 14 public declarations |
+| A+ | 100 | 144 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionRuleTable.swift` | no automated issues |
+| A+ | 100 | 158 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/PermissionWildcardPattern.swift` | no automated issues |
 | A+ | 97 | 271 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/Safety.swift` | 21 public declarations |
 | A+ | 100 | 292 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/StaticSafetyDownloadPolicy.swift` | no automated issues |
 | A+ | 100 | 229 | `source:QuillCodeSafety` | `Sources/QuillCodeSafety/StaticSafetyPolicy.swift` | no automated issues |
